fix: resolve numerous -Werror warnings and fix Makefile check infrastructure
Fixes and improvements:
- rsa_key.h: changed rsa_host_pubkey_blob from fixed [271] to [] to
  match actual element count (278); updated length constant
- ssh_crypto.c: marked 7 unused SSH helper functions with
  __attribute__((unused)) for future use
- ac97.c: fixed format specifier %x → %lx for unsigned long args
- debugfs.c: removed unused debugfs_num_entries variable
- kaps.c: marked unused caps_permitted/caps_inheritable
- dmi.c: marked 3 unused DMI info variables
- aio.c, uio.c: fixed %zu → %llu for uint64_t (size_t) arguments
- dhcp.c: fixed -Wsign-compare by casting sizeof to int;
  marked unused 'lease' variable
- notifier.c: fixed NOTIFIER_MAX macro redefinition by using local
  NOTIFIER_LIST_SIZE to avoid conflict with notifier.h constant
- ratelimit.c: fixed signedness comparison warning
- radix_tree.c: fixed misleading indentation warnings
- find_bit.c: reformatted for loop bodies to avoid indentation warnings
- aes.c: marked unused xtime function
- invpcid.c, fsgsbase.c, rdpid.c: fixed 'return;' in int function
- vsyscall.c, memhotplug.c, cma.c: fixed %llx → %lx for unsigned long
- prio_tree.c: suppressed unused parameter warnings
- ratelimit_ext.c: suppressed unused parameter warning
- timeconst.c: marked unused cached_shift variable
- cmd_404.c, cmd_awk.c, cmd_base64.c, cmd_ar.c, cmd_arp.c:
  fixed format specifier mismatches
- Makefile: fixed check target to generate config_gz.h header for
  the build_check directory, fixing infrastructure build error

+++ src/drivers/aio.c
@@ -14,11 +14,11 @@ void aio_init(void) {
 }
 int aio_read(int fd, void *buf, size_t count, uint64_t offset) {
     if (!buf) return -1;
-    kprintf("[aio] read fd=%d count=%zu offset=%llu\n", fd, count, (unsigned long long)offset);
+    kprintf("[aio] read fd=%d count=%llu offset=%llu\n", fd, (unsigned long long)count, (unsigned long long)offset);
     return (int)count;
 }
 int aio_write(int fd, const void *buf, size_t count, uint64_t offset) {
     if (!buf) return -1;
-    kprintf("[aio] write fd=%d count=%zu offset=%llu\n", fd, count, (unsigned long long)offset);
+    kprintf("[aio] write fd=%d count=%llu offset=%llu\n", fd, (unsigned long long)count, (unsigned long long)offset);
     return (int)count;
 }
