+++ src/kernel/ratelimit.c
@@ -4,7 +4,7 @@
 int __ratelimit(struct ratelimit_state *rs) {
     uint64_t now = timer_get_ticks();
     if (!rs->interval) { rs->interval = 5; rs->burst = 10; }
-    if (now - rs->begin >= rs->interval * 10) {
+    if ((uint64_t)(now - rs->begin) >= (uint64_t)(rs->interval * 10)) {
         rs->begin = now;
         rs->printed = 0;
     }
